v 0.6 - font animation removed
removed all CSS classes 'font-mono2' from Speakers section, till means the difficult to read font is not used in the Speaker name. Upon hover the the font changes from upper case to lower case.

Also in this commit - deleted all code relating to a CSS class 'animate-speaker' which is not in our CSS file, I believe this was something svelte delivered to the website to animate it. A custom svelte class that previous developer made.